requirements.txt
pandas
scikit-learn
tqdm
pyyaml
levenshtein
torch==2.3.1
transformers==4.41.2
datasets==2.20.0
bitsandbytes==0.43.1
peft==0.11.1
trl==0.9.4